[DOC] remove empty boxes in README (#683)
PR https://github.com/PyPortfolio/PyPortfolioOpt/pull/676 added empty
boxes in the readme used for testing.

This is confusing for the user, as the empty blocks get displayed as
empty boxes, creating the impression of an error or misrendering.

This needs to be removed, and we should think separately about how the
testing can be achieved nonetheless.

FYI @tschm